future/repos/Faceless_Shorts/requirements.txt
# Faceless Shorts Automator MVP – Python dependencies
# Install: pip install -r requirements.txt

python-dotenv>=1.0.0
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
elevenlabs>=1.0.0
gTTS>=2.4.0
Pillow>=10.0.0
moviepy>=1.0.3
